+++ src/CfdiUtils/CadenaOrigen/XsltBuildException.php
@@ -1,0 +1,6 @@
+<?php
+namespace CfdiUtils\CadenaOrigen;
+
+class XsltBuildException extends \RuntimeException
+{
+}