+++ src/CfdiUtils/CadenaOrigen/XsltBuilderPropertyInterface.php
@@ -1,0 +1,8 @@
+<?php
+namespace CfdiUtils\CadenaOrigen;
+
+interface XsltBuilderPropertyInterface
+{
+    public function getXsltBuilder(): XsltBuilderInterface;
+    public function setXsltBuilder(XsltBuilderInterface $xsltBuilder);
+}