+++ src/CfdiUtils/Validate/Contracts/RequireXsltBuilderInterface.php
@@ -1,0 +1,8 @@
+<?php
+namespace CfdiUtils\Validate\Contracts;
+
+use CfdiUtils\CadenaOrigen\XsltBuilderPropertyInterface;
+
+interface RequireXsltBuilderInterface extends XsltBuilderPropertyInterface
+{
+}