+++ src/CfdiUtils/Validate/Hydrater.php
@@ -1,17 +1,21 @@
 <?php
 namespace CfdiUtils\Validate;
 
+use CfdiUtils\CadenaOrigen\XsltBuilderPropertyInterface;
+use CfdiUtils\CadenaOrigen\XsltBuilderPropertyTrait;
 use CfdiUtils\Validate\Contracts\RequireXmlResolverInterface;
 use CfdiUtils\Validate\Contracts\RequireXmlStringInterface;
+use CfdiUtils\Validate\Contracts\RequireXsltBuilderInterface;
 use CfdiUtils\Validate\Contracts\ValidatorInterface;
 use CfdiUtils\Validate\Traits\XmlStringPropertyTrait;
 use CfdiUtils\XmlResolver\XmlResolverPropertyInterface;
 use CfdiUtils\XmlResolver\XmlResolverPropertyTrait;
 
-class Hydrater implements XmlResolverPropertyInterface
+class Hydrater implements XmlResolverPropertyInterface, XsltBuilderPropertyInterface
 {
     use XmlResolverPropertyTrait;
     use XmlStringPropertyTrait;
+    use XsltBuilderPropertyTrait;
 
     public function hydrate(ValidatorInterface $validator)
     {
@@ -20,6 +24,9 @@ class Hydrater implements XmlResolverPropertyInterface
         }
         if ($this->hasXmlResolver() && $validator instanceof RequireXmlResolverInterface) {
             $validator->setXmlResolver($this->getXmlResolver());
+        }
+        if ($validator instanceof RequireXsltBuilderInterface) {
+            $validator->setXsltBuilder($this->getXsltBuilder());
         }
     }
 }